Returns an error message > 2048 and channel is a voice channel

+++ commands/config/configjoin.js
@@ -27,6 +27,7 @@ module.exports = class extends Command {
         let collected = await message.channel.awaitMessages(filter, opt).catch(() => {});
         if (!collected || !collected.first()) return msg.error("common:CANCELLED", null, true);
         const confMessage = collected.first().content;
+        if (confMessage > 2048) return msg.error("config/configjoin:CARACTERES2048", null, true);
         if (confMessage === "cancel") return msg.error("common:CANCELLED", null, true);
         if (confMessage === data.guild.prefix+"setjoin") return;
         collected.first().delete();
@@ -40,7 +41,7 @@ module.exports = class extends Command {
         const channel = confChannel.mentions.channels.first()
         || message.guild.channels.cache.get(confChannel.content)
         || message.guild.channels.cache.find((ch) => ch.name === confChannel.content || `#${ch.name}` === confChannel.content);
-        if (!channel) return msg.error("config/configjoin:CHANNEL_NOT_FOUND", {
+        if (!channel || channel.type === "voice") return msg.error("config/configjoin:CHANNEL_NOT_FOUND", {
             channel: confChannel.content
         }, true);
         collected.first().delete();
